+++ CHANGELOG.rst
@@ -28,6 +28,8 @@ Changed
   deprecated)
 - persist action names in domain during model persistence
 - improved travis build speed by not using miniconda
+- don't fail with an exception but with a helpful error message if an
+  utterance template contains a variable that can not be filled
 
 Removed
 -------